requirements.txt
asgiref==3.4.1
Django==3.2.9
django-ckeditor==6.2.0
django-environ==0.8.1
django-js-asset==1.2.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
uuid==1.30
